third_party/blink/web_tests/external/wpt/editing/run/fontsize_2001-last-expected.txt

This is a testharness.js-based test.
Found 9 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "foo<font size=2>ba[r</font>b]az" queryCommandState("stylewithcss") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "foo<font size=2>ba[r</font>b]az" queryCommandIndeterm("fontsize") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "foo<font size=2>ba[r</font>b]az" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "2"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "fo[o<font size=2>bar</font>b]az" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "fo[o<font size=2>bar</font>b]az" queryCommandIndeterm("fontsize") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "fo[o<font size=2>bar</font>b]az" queryCommandIndeterm("fontsize") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<font size=1>fo[o</font><span style=font-size:xx-small>b]ar</span>" queryCommandIndeterm("fontsize") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "<font size=1>fo[o</font><span style=font-size:xx-small>b]ar</span>" queryCommandIndeterm("fontsize") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["fontsize","4"]] "<font size=4>fo[o</font><span style=font-size:large>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<font size=\\"4\\">foo</font><span style=\\"font-size:large\\">bar</span>" but got "<font size=\\"4\\">foob</font><span style=\\"font-size:large\\">ar</span>"
Harness: the test ran to completion.
